MidiFile->BeatsPerMinute into double. Accordingly also the MidiEvent.cs->DeltaTimes and MidiTrack.cs->TotalTime had to be changed.
Changed in SampleSequencer the parameter from uint to double.
private double DeltaTimetoSamples(double DeltaTime)

MidiSequencer synchronized with audio in Unity works pretty nice without any lags.

Removed all possible warnings from the project.

+++ branches/V1.0/src/CSharpSynth/Midi/MidiEvent.cs
@@ -3,7 +3,7 @@
     public class MidiEvent
     {
         //--Variables
-        public uint deltaTime;
+        public double deltaTime;
         public MidiHelper.MidiMetaEvent midiMetaEvent;
         public MidiHelper.MidiChannelEvent midiChannelEvent;
         public object[] Parameters;
